pyproject.toml
[tool.poetry]
name = "automated-tending-machine"
version = "0.0.1"
description = ""
authors = ["Ray Andrew <raydreww@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.5"
briefcase = "^0.2.10"
PySide2 = "^5.13.1"
qtpy = "^1.9.0"
gpiozero = "^1.5"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.1"
pylint = "^2.4"
flake8 = "^3.7"
pytest-cov = "^2.8"
pytest-qt = "^3.2"
pytest-xdist = "^1.30"
pytest-mock = "^1.11"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
